// src/app/api/generated/model/geographicalData.ts
/**
 * BE
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GeographicalData { 
    id?: number;
    openbareruimte?: string | null;
    huisnummer?: number;
    huisletter?: string | null;
    huisnummertoevoeging?: number | null;
    postcode?: string | null;
    woonplaats?: string | null;
    gemeente?: string | null;
    provincie?: string | null;
    nummeraanduiding?: string | null;
    verblijfsobjectgebruiksdoel?: string | null;
    oppervlakteverblijfsobject?: number;
    verblijfsobjectstatus?: string | null;
    objectId?: string | null;
    objectType?: string | null;
    nevenadres?: string | null;
    pandid?: string | null;
    pandstatus?: string | null;
    pandbouwjaar?: number;
    x?: number;
    y?: number;
    lon?: number;
    lat?: number;
}