// src/app/api/generated/model/updateGeographicalDataDto.ts
/**
 * Geographical Data API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateGeographicalDataDto { 
    openbareruimte: string;
    huisnummer: number;
    huisletter?: string | null;
    huisnummertoevoeging?: number | null;
    postcode: string;
    woonplaats: string;
    gemeente?: string | null;
    provincie?: string | null;
    nummeraanduiding?: string | null;
    verblijfsobjectgebruiksdoel?: string | null;
    oppervlakteverblijfsobject?: number;
    verblijfsobjectstatus?: string | null;
    objectId?: string | null;
    objectType?: string | null;
    nevenadres?: string | null;
    pandid?: string | null;
    pandstatus?: string | null;
    pandbouwjaar?: number;
    x?: number;
    y?: number;
    lon?: number;
    lat?: number;
    id?: number;
}